fix(gus): ramp increment is in significand units — ours ran 16x too slow
The GF1 ramp start/end registers hold the TOP BYTE of the 16-bit log volume
(`reg << 8`), but the rate register's 6-bit increment does not count in those
units: it steps the volume's 12-bit significand, whose LSB is bit 4 of the
16-bit word. We fed the raw field straight through, so every volume ramp took
16x as long as the guest asked for — a ~1.5ms note attack became ~24ms.

That is not a subtle detune, because DMX polls the ramp-done bit (GF1 0x8D)
with interrupts disabled: the guest sat in that spin long enough to miss its
own 140Hz music ticks, and the vPIC coalesces a re-raised edge into the
still-pending IRR bit, so each missed tick is one the sequencer never counts.
Emulated port traps over a Doom run: ~200k/s -> ~8k/s, and the GF1 poll storm
(50-90k reads/s of one register) disappears entirely.

Scale the field at the GF1 front-end (`ramp_inc`) so `Ramp::inc` is always in
the sampler's own log-volume domain, same as floor/ceil/vol — the sampler is
shared with the coming MIDI personalities and should not carry a GF1 register
encoding.

Also lands the mixer-pump work this was found under: every device renders into
one fixed 44.1kHz canonical clock and resamples itself (letting the 11kHz DSP
own the rate was band-limiting the GUS music to 5.5kHz), the DSP gets a local
frame epoch so an SB effect starting over GUS music no longer re-keys the GUS
event timeline, and single-cycle DMA completes on its own device clock instead
of waiting behind the shared sink's cushion.

+++ lib/sampler/src/voice.rs
@@ -28,8 +28,14 @@ pub enum LoopMode {
 #[derive(Clone, Copy, Default)]
 pub struct Ramp {
     pub running: bool,
-    /// Amount added per update (the GF1 ramp-rate register's bits 5..0).
-    pub inc: u8,
+    /// Amount added per update, in the SAME log-volume domain as `floor` /
+    /// `ceil` / `Voice::vol`. Not the GF1 rate register's raw 6-bit field:
+    /// that counts in units of the volume's 12-bit significand (bit 4 of the
+    /// 16-bit register), so the GF1 front-end scales it by 16 on the way in.
+    /// Feeding the raw field straight through made every ramp 16x too slow —
+    /// a ~1.5 ms note attack took ~24 ms, and DMX busy-waits (interrupts off)
+    /// for the ramp to finish, which swallowed the guest's 140 Hz music ticks.
+    pub inc: u16,
     /// Update-rate divider (rate bits 7..6): apply every `1 << (3*shift)`
     /// frames — 1, 8, 64, 512.
     pub shift: u8,
